Renamed train_model_1.py and train_model_1_launch.py to train_model.py and train_model_launch.py
