+++ train_model_launch.py
@@ -7,4 +7,4 @@ from config import *
 
 if __name__ == "__main__":
 	run_name = MODEL_TYPE
-	os.system('SGE_Batch -q gpu.q -r "{}" -c "python3 -u train_model_1.py {}" -P {}'.format(run_name, False, JOB_PRIORITY))
+	os.system('SGE_Batch -q gpu.q -r "{}" -c "python3 -u train_model.py {}" -P {}'.format(run_name, False, JOB_PRIORITY))
